+++ src/Components/AddExpense/AddExpense.css
@@ -1,6 +1,6 @@
 .add-expense-area {
     padding: 10rem 2rem;
-    background-image: linear-gradient(to bottom, rgba(255, 209, 102, 0.7), rgba(255, 109, 102, 0.3))  , url("../../res/add-expense-background.jpg") ;
+    background-image: linear-gradient(to bottom, rgba(255, 209, 102, 0.7), rgba(255, 109, 102, 0.3)), url("../../res/add-expense-background.jpg");
     background-position: center;
     background-repeat: no-repeat;
     background-size: cover;
@@ -11,10 +11,29 @@
     padding: 1rem;
     margin: 2rem auto;
     width: 50rem;
+    height: 50rem;
     max-width: 95%;
     border-radius: 12px;
     text-align: center;
     box-shadow: 0 1px 8px rgba(0, 0, 0, 0.25);
+    display: flex;
+    flex-flow: column nowrap;
+    justify-content: center;
+    align-items: flex-start;
+}
+
+.column-align-start {
+    display: flex;
+    flex-flow: column nowrap;
+    justify-content: center;
+    align-items: flex-start;
+}
+
+.column-align-center {
+    display: flex;
+    flex-flow: column nowrap;
+    justify-content: center;
+    align-items: center;
 }
 
 .add-expense button {
@@ -46,3 +65,14 @@
     background-color: var(--color-yellow-0);
 }
 
+@media screen and (max-width: 500px) {
+    .add-expense {
+        height: 60rem;
+        align-items: center;
+
+    }
+
+    .add-expense button:not(:last-of-type) {
+        margin-bottom: 2rem;
+    }
+}
