+++ src/Components/ExpenseForm/ExpenseForm.css
@@ -1,3 +1,7 @@
+.add-expense-form {
+    width: 100%;
+}
+
 .add-expense__title {
     margin-bottom: 5rem;
 }
@@ -6,7 +10,6 @@
     display: flex;
     flex-flow: column nowrap;
     gap: 1rem;
-    margin-bottom: 1rem;
     text-align: left;
     padding: 5rem 3rem;
 }
@@ -33,4 +36,16 @@
 
 .add-expense__actions {
     text-align: right;
+}
+
+@media screen and (max-width: 500px) {
+
+    .add-expense__controls {
+        padding: 4rem 2rem;
+
+    }
+
+    .add-expense__control input {
+        margin: 0px auto;
+    }
 }
