+++ src/Container/Header/Header.jsx
@@ -1,4 +1,4 @@
-import React, {useEffect} from 'react';
+import React from 'react';
 import "./Header.css"
 
 function Header(props) {
